+++ src/app/shared/components/input-datapicker/input-datapicker.component.spec.ts
@@ -1,0 +1,24 @@
+import { ComponentFixture, TestBed, waitForAsync } from "@angular/core/testing";
+import { IonicModule } from "@ionic/angular";
+
+import { InputDatapickerComponent } from "./input-datapicker.component";
+
+describe("InputDatapickerComponent", () => {
+  let component: InputDatapickerComponent;
+  let fixture: ComponentFixture<InputDatapickerComponent>;
+
+  beforeEach(waitForAsync(() => {
+    TestBed.configureTestingModule({
+      declarations: [InputDatapickerComponent],
+      imports: [IonicModule.forRoot()],
+    }).compileComponents();
+
+    fixture = TestBed.createComponent(InputDatapickerComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  }));
+
+  it("should create", () => {
+    expect(component).toBeTruthy();
+  });
+});